Creacion de clases usuario y rol para comenzar con el registro de usuario, cambio de declaracion de la variable "id" de la clase producto de "Integer" por "Long"

+++ src/main/java/com/RestauranteSpring/model/Producto.java
@@ -12,12 +12,12 @@ public class Producto {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String nombre;
 	private String descripcion;
 	private Double precio;
 
-	public Producto(Integer id, String nombre, String descripcion, Double precio) {
+	public Producto(Long id, String nombre, String descripcion, Double precio) {
 		super();
 		this.id = id;
 		this.nombre = nombre;
@@ -29,11 +29,11 @@ public class Producto {
 		super();
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
